Moved ViewModel into its own file.
