--- app/src/main/java/com/example/watcomtravels/BasicMap.kt
@@ -1,2 +1,0 @@
-package com.example.watcomtravels
-
